Enum ic_system_api::ApiType

pub enum ApiType {
    Start,
    Init {
        time: Time,
        incoming_payload: Vec<u8>,
        caller: PrincipalId,
    },
    Update {
        time: Time,
        incoming_payload: Vec<u8>,
        incoming_cycles: Cycles,
        caller: PrincipalId,
        call_context_id: CallContextId,
        response_data: Vec<u8>,
        response_status: ResponseStatus,
        own_subnet_id: SubnetId,
        own_subnet_type: SubnetType,
        routing_table: Arc<RoutingTable>,
        subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>,
        outgoing_request: Option<RequestInPrep>,
        max_reply_size: NumBytes,
    },
    ReplicatedQuery {
        time: Time,
        incoming_payload: Vec<u8>,
        caller: PrincipalId,
        response_data: Vec<u8>,
        response_status: ResponseStatus,
        data_certificate: Option<Vec<u8>>,
        max_reply_size: NumBytes,
    },
    NonReplicatedQuery {
        time: Time,
        incoming_payload: Vec<u8>,
        caller: PrincipalId,
        call_context_id: CallContextId,
        data_certificate: Option<Vec<u8>>,
        own_subnet_id: SubnetId,
        routing_table: Arc<RoutingTable>,
        outgoing_request: Option<RequestInPrep>,
        response_data: Vec<u8>,
        response_status: ResponseStatus,
        max_reply_size: NumBytes,
    },
    ReplyCallback {
        time: Time,
        incoming_payload: Vec<u8>,
        incoming_cycles: Cycles,
        call_context_id: CallContextId,
        response_data: Vec<u8>,
        response_status: ResponseStatus,
        own_subnet_id: SubnetId,
        own_subnet_type: SubnetType,
        routing_table: Arc<RoutingTable>,
        subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>,
        outgoing_request: Option<RequestInPrep>,
        max_reply_size: NumBytes,
    },
    RejectCallback {
        time: Time,
        reject_context: RejectContext,
        incoming_cycles: Cycles,
        call_context_id: CallContextId,
        response_data: Vec<u8>,
        response_status: ResponseStatus,
        own_subnet_id: SubnetId,
        own_subnet_type: SubnetType,
        routing_table: Arc<RoutingTable>,
        subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>,
        outgoing_request: Option<RequestInPrep>,
        max_reply_size: NumBytes,
    },
    PreUpgrade {
        caller: PrincipalId,
        time: Time,
    },
    InspectMessage {
        caller: PrincipalId,
        method_name: String,
        incoming_payload: Vec<u8>,
        time: Time,
        message_accepted: bool,
    },
    Heartbeat {
        time: Time,
        call_context_id: CallContextId,
        own_subnet_id: SubnetId,
        own_subnet_type: SubnetType,
        routing_table: Arc<RoutingTable>,
        subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>,
        outgoing_request: Option<RequestInPrep>,
    },
    Cleanup {
        time: Time,
    },
}

Variants

Start

Init

Fields of Init

time: Timeincoming_payload: Vec<u8>caller: PrincipalId

Update

Fields of Update

time: Timeincoming_payload: Vec<u8>incoming_cycles: Cyclescaller: PrincipalIdcall_context_id: CallContextIdresponse_data: Vec<u8>response_status: ResponseStatusown_subnet_id: SubnetIdown_subnet_type: SubnetTyperouting_table: Arc<RoutingTable>subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>outgoing_request: Option<RequestInPrep>

Optional outgoing request under construction. If None no outgoing request is currently under construction.

max_reply_size: NumBytes

ReplicatedQuery

Fields of ReplicatedQuery

time: Timeincoming_payload: Vec<u8>caller: PrincipalIdresponse_data: Vec<u8>response_status: ResponseStatusdata_certificate: Option<Vec<u8>>max_reply_size: NumBytes

NonReplicatedQuery

Fields of NonReplicatedQuery

time: Timeincoming_payload: Vec<u8>caller: PrincipalIdcall_context_id: CallContextIddata_certificate: Option<Vec<u8>>own_subnet_id: SubnetIdrouting_table: Arc<RoutingTable>outgoing_request: Option<RequestInPrep>

Optional outgoing request under construction. If None no outgoing request is currently under construction.

response_data: Vec<u8>response_status: ResponseStatusmax_reply_size: NumBytes

ReplyCallback

Fields of ReplyCallback

time: Timeincoming_payload: Vec<u8>incoming_cycles: Cyclescall_context_id: CallContextIdresponse_data: Vec<u8>response_status: ResponseStatusown_subnet_id: SubnetIdown_subnet_type: SubnetTyperouting_table: Arc<RoutingTable>subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>outgoing_request: Option<RequestInPrep>

Optional outgoing request under construction. If None no outgoing request is currently under construction.

max_reply_size: NumBytes

RejectCallback

Fields of RejectCallback

time: Timereject_context: RejectContextincoming_cycles: Cyclescall_context_id: CallContextIdresponse_data: Vec<u8>response_status: ResponseStatusown_subnet_id: SubnetIdown_subnet_type: SubnetTyperouting_table: Arc<RoutingTable>subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>outgoing_request: Option<RequestInPrep>

Optional outgoing request under construction. If None no outgoing request is currently under construction.

max_reply_size: NumBytes

PreUpgrade

Fields of PreUpgrade

caller: PrincipalIdtime: Time

InspectMessage

For executing canister_inspect_message method that allows the canister to decide pre-consensus if it actually wants to accept the message or not.

Fields of InspectMessage

caller: PrincipalIdmethod_name: Stringincoming_payload: Vec<u8>time: Timemessage_accepted: bool

Heartbeat

Fields of Heartbeat

time: Timecall_context_id: CallContextIdown_subnet_id: SubnetIdown_subnet_type: SubnetTyperouting_table: Arc<RoutingTable>subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>outgoing_request: Option<RequestInPrep>

Optional outgoing request under construction. If None no outgoing request is currently under construction.

Cleanup

For executing the call_on_cleanup callback.

The call_on_cleanup callback is executed iff the reply or the reject callback was executed and trapped (for any reason).

See https://sdk.dfinity.org/docs/interface-spec/index.html#system-api-call

Fields of Cleanup

time: Time

Implementations

impl ApiType

pub fn start() -> Self

pub fn init(time: Time, incoming_payload: Vec<u8>, caller: PrincipalId) -> Self

pub fn heartbeat(
time: Time,
call_context_id: CallContextId,
own_subnet_id: SubnetId,
own_subnet_type: SubnetType,
routing_table: Arc<RoutingTable>,
subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>
) -> Self

pub fn update(
time: Time,
incoming_payload: Vec<u8>,
incoming_cycles: Cycles,
caller: PrincipalId,
call_context_id: CallContextId,
own_subnet_id: SubnetId,
own_subnet_type: SubnetType,
routing_table: Arc<RoutingTable>,
subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>
) -> Self

pub fn replicated_query(
time: Time,
incoming_payload: Vec<u8>,
caller: PrincipalId,
data_certificate: Option<Vec<u8>>
) -> Self

pub fn non_replicated_query(
time: Time,
incoming_payload: Vec<u8>,
caller: PrincipalId,
call_context_id: CallContextId,
own_subnet_id: SubnetId,
routing_table: Arc<RoutingTable>,
data_certificate: Option<Vec<u8>>
) -> Self

pub fn reply_callback(
time: Time,
incoming_payload: Vec<u8>,
incoming_cycles: Cycles,
call_context_id: CallContextId,
replied: bool,
own_subnet_id: SubnetId,
own_subnet_type: SubnetType,
routing_table: Arc<RoutingTable>,
subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>
) -> Self

pub fn reject_callback(
time: Time,
reject_context: RejectContext,
incoming_cycles: Cycles,
call_context_id: CallContextId,
replied: bool,
own_subnet_id: SubnetId,
own_subnet_type: SubnetType,
routing_table: Arc<RoutingTable>,
subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>
) -> Self

pub fn pre_upgrade(time: Time, caller: PrincipalId) -> Self

pub fn inspect_message(
caller: PrincipalId,
method_name: String,
incoming_payload: Vec<u8>,
time: Time
) -> Self