Trait ic_system_api::SystemStateAccessor

pub trait SystemStateAccessor {
    fn canister_id(&self) -> CanisterId;

    fn controller(&self) -> PrincipalId;

    fn mint_cycles(&self, amount: Cycles) -> HypervisorResult<()>;

    fn msg_cycles_accept(
        &self, 
        call_context_id: &CallContextId, 
        max_amount: Cycles
    ) -> Cycles;

    fn msg_cycles_available(
        &self, 
        call_context_id: &CallContextId
    ) -> HypervisorResult<Cycles>;

    fn stable_size(&self) -> u32;

    fn stable_grow(&self, additional_pages: u32) -> i32;

    fn stable_read(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> Result<(), StableMemoryError>;

    fn stable_write(
        &self, 
        offset: u32, 
        src: u32, 
        size: u32, 
        heap: &[u8]
    ) -> Result<(), StableMemoryError>;

    fn canister_cycles_balance(&self) -> Cycles;

    fn canister_cycles_withdraw(
        &self, 
        canister_current_memory_usage: NumBytes, 
        canister_compute_allocation: ComputeAllocation, 
        amount: Cycles
    ) -> HypervisorResult<()>;

    fn canister_cycles_refund(&self, cycles: Cycles);

    fn set_certified_data(&self, data: Vec<u8>);

    fn register_callback(&self, callback: Callback) -> CallbackId;

    fn unregister_callback(&self, callback_id: CallbackId) -> Option<Callback>;

    fn push_output_request(
        &self, 
        canister_current_memory_usage: NumBytes, 
        canister_compute_allocation: ComputeAllocation, 
        msg: Request
    ) -> Result<(), (StateError, Request)>;

    fn canister_status(&self) -> CanisterStatus;
}

The abstract interface through which canister user code can affect the canister system state. This layer is a slightly higher level of abstraction than "raw" system call code -- it is assumed that system call code has already resolved/demarshalled raw syscall arguments.

Required methods

fn canister_id(&self) -> CanisterId

Obtains the canister ID.

fn controller(&self) -> PrincipalId

Obtains the controller of this canister.

fn mint_cycles(&self, amount: Cycles) -> HypervisorResult<()>

Increases the balance of the canister by amount

fn msg_cycles_accept(
&self,
call_context_id: &CallContextId,
max_amount: Cycles
) -> Cycles

Accepts cycles from given call context.

fn msg_cycles_available(
&self,
call_context_id: &CallContextId
) -> HypervisorResult<Cycles>

Determines cycles given in call context.

fn stable_size(&self) -> u32

Determines size of stable memory in Web assembly pages.

fn stable_grow(&self, additional_pages: u32) -> i32

Grows stable memory by specified amount.

fn stable_read(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> Result<(), StableMemoryError>

Reads from stable memory back to heap.

fn stable_write(
&self,
offset: u32,
src: u32,
size: u32,
heap: &[u8]
) -> Result<(), StableMemoryError>

Writes from heap to stable memory.

fn canister_cycles_balance(&self) -> Cycles

Current cycles balance of the canister.

fn canister_cycles_withdraw(
&self,
canister_current_memory_usage: NumBytes,
canister_compute_allocation: ComputeAllocation,
amount: Cycles
) -> HypervisorResult<()>

Withdraws cycles from canister's balance.

fn canister_cycles_refund(&self, cycles: Cycles)

(Re-)add cycles to canister. This is intended to be used to reclaim cycles from unfulfilled requests.

fn set_certified_data(&self, data: Vec<u8>)

Set certified data.

fn register_callback(&self, callback: Callback) -> CallbackId

Registers callback for call return.

fn unregister_callback(&self, callback_id: CallbackId) -> Option<Callback>

Unregister callback for call return.

fn push_output_request(
&self,
canister_current_memory_usage: NumBytes,
canister_compute_allocation: ComputeAllocation,
msg: Request
) -> Result<(), (StateError, Request)>

Pushes outgoing request.

fn canister_status(&self) -> CanisterStatus

Current status of canister.

Loading content...